Move testing.py to tests folder
